WIP: Settings links   - Hyperlinks for terms of use, privacy, and learn loop
