+++ LoopTripIOS/Assets.xcassets/ICO Trash.imageset/Contents.json
