+++ LoopTripIOS/Assets.xcassets/ICO Trash.imageset/ICO Trash Off.png
