+++ LoopTripIOS/Assets.xcassets/ICO Trash.imageset/ICO Trash Off@2x.png
