+++ LoopTripIOS/Assets.xcassets/ICO Trash.imageset/ICO Trash Off@3x.png
